+++ CharacterAlarm/Repository/APIRepository.swift
@@ -10,7 +10,7 @@ extension APIRepository {
         let url = URL(string: environmentVariable.apiEndpoint + path)!
         let requestHeader: [String: String] = APIHeader.createAuthorizationRequestHeader(userID: userID, authToken: authToken)
         let requestBody: Encodable = pushToken
-        let messageResponse: MessageResponse = try await APIClient2().request(url: url, httpMethod: .post, requestHeader: requestHeader, requestBody: requestBody)
+        let messageResponse: MessageResponse = try await APIClient().request(url: url, httpMethod: .post, requestHeader: requestHeader, requestBody: requestBody)
     }
 
     func postPushTokenAddVoIPPushToken(userID: String, authToken: String, pushToken: PushTokenRequest) async throws {
@@ -18,7 +18,7 @@ extension APIRepository {
         let url = URL(string: environmentVariable.apiEndpoint + path)!
         let requestHeader: [String: String] = APIHeader.createAuthorizationRequestHeader(userID: userID, authToken: authToken)
         let requestBody: Encodable = pushToken
-        let messageResponse: MessageResponse = try await APIClient2().request(url: url, httpMethod: .post, requestHeader: requestHeader, requestBody: requestBody)
+        let messageResponse: MessageResponse = try await APIClient().request(url: url, httpMethod: .post, requestHeader: requestHeader, requestBody: requestBody)
     }
 }
 
@@ -28,7 +28,7 @@ extension APIRepository {
         let url = URL(string: environmentVariable.apiEndpoint + path)!
         let requestHeader = APIHeader.createAuthorizationRequestHeader(userID: userID, authToken: authToken)
         let requestBody: Encodable? = nil
-        let userInfoResponse: UserInfoResponse = try await APIClient2().request(url: url, httpMethod: .post, requestHeader: requestHeader, requestBody: requestBody)
+        let userInfoResponse: UserInfoResponse = try await APIClient().request(url: url, httpMethod: .post, requestHeader: requestHeader, requestBody: requestBody)
         
         // 変換
         let iOSPlatformInfo = UserInfoIOSPlatformInfo (
@@ -49,7 +49,7 @@ extension APIRepository {
         let url = URL(string: environmentVariable.apiEndpoint + path)!
         let requestHeader: [String: String] = APIHeader.defaultHeader
         let requestBody: Encodable? = request
-        let messageResponse: MessageResponse = try await APIClient2().request(url: url, httpMethod: .post, requestHeader: requestHeader, requestBody: requestBody)
+        let messageResponse: MessageResponse = try await APIClient().request(url: url, httpMethod: .post, requestHeader: requestHeader, requestBody: requestBody)
     }
     
     func postUserWithdraw(userID: String, authToken: String) async throws {
@@ -57,7 +57,7 @@ extension APIRepository {
         let url = URL(string: environmentVariable.apiEndpoint + path)!
         let requestHeader: [String: String] = APIHeader.createAuthorizationRequestHeader(userID: userID, authToken: authToken)
         let requestBody: Encodable? = nil
-        let messageResponse: MessageResponse = try await APIClient2().request(url: url, httpMethod: .post, requestHeader: requestHeader, requestBody: requestBody)
+        let messageResponse: MessageResponse = try await APIClient().request(url: url, httpMethod: .post, requestHeader: requestHeader, requestBody: requestBody)
     }
 }
 
@@ -67,7 +67,7 @@ extension APIRepository {
         let url = URL(string: environmentVariable.apiEndpoint + path)!
         let requestHeader: [String: String] = APIHeader.defaultHeader
         let requestBody: Encodable? = nil
-        let charaResponses: [CharaResponse] = try await APIClient2().request(url: url, httpMethod: .get, requestHeader: requestHeader, requestBody: requestBody)
+        let charaResponses: [CharaResponse] = try await APIClient().request(url: url, httpMethod: .get, requestHeader: requestHeader, requestBody: requestBody)
         return charaResponses.map { Chara(charaResponse: $0) }
     }
     
@@ -76,7 +76,7 @@ extension APIRepository {
         let url = URL(string: environmentVariable.apiEndpoint + path)!
         let requestHeader: [String: String] = APIHeader.defaultHeader
         let requestBody: Encodable? = nil
-        let charaResponse: CharaResponse = try await APIClient2().request(url: url, httpMethod: .get, requestHeader: requestHeader, requestBody: requestBody)
+        let charaResponse: CharaResponse = try await APIClient().request(url: url, httpMethod: .get, requestHeader: requestHeader, requestBody: requestBody)
         return Chara(charaResponse: charaResponse)
     }
 }
@@ -87,7 +87,7 @@ extension APIRepository {
         let url = URL(string: environmentVariable.apiEndpoint + path)!
         let requestHeader = APIHeader.defaultHeader
         let requestBody: Request? = nil
-        let response: MaintenanceResponse = try await APIClient2().request(url: url, httpMethod: .get, requestHeader: requestHeader, requestBody: requestBody)
+        let response: MaintenanceResponse = try await APIClient().request(url: url, httpMethod: .get, requestHeader: requestHeader, requestBody: requestBody)
         return response.maintenance
     }
     
@@ -96,7 +96,42 @@ extension APIRepository {
         let url = URL(string: environmentVariable.apiEndpoint + path)!
         let requestHeader = APIHeader.defaultHeader
         let requestBody: Request? = nil
-        let response: RequireVersionResponse = try await APIClient2().request(url: url, httpMethod: .get, requestHeader: requestHeader, requestBody: requestBody)
+        let response: RequireVersionResponse = try await APIClient().request(url: url, httpMethod: .get, requestHeader: requestHeader, requestBody: requestBody)
         return response.iosVersion
+    }
+}
+
+
+extension APIRepository {
+    func fetchAlarms(userID: String, authToken: String) async throws -> [AlarmResponse] {
+        let path = "/alarm/list"
+        let url = URL(string: environmentVariable.apiEndpoint + path)!
+        let requestHeader = APIHeader.createAuthorizationRequestHeader(userID: userID, authToken: authToken)
+        let requestBody: Request? = nil
+        return try await APIClient().request(url: url, httpMethod: .post, requestHeader: requestHeader, requestBody: requestBody)
+    }
+    
+    func addAlarm(userID: String, authToken: String, requestBody: AlarmAddRequest) async throws {
+        let path = "/alarm/add"
+        let url = URL(string: environmentVariable.apiEndpoint + path)!
+        let requestHeader = APIHeader.createAuthorizationRequestHeader(userID: userID, authToken: authToken)
+        let requestBody: Request? = requestBody
+        let response: MessageResponse = try await APIClient().request(url: url, httpMethod: .post, requestHeader: requestHeader, requestBody: requestBody)
+    }
+
+    func editAlarm(userID: String, authToken: String, requestBody: AlarmEditRequest) async throws {
+        let path = "/alarm/edit"
+        let url = URL(string: environmentVariable.apiEndpoint + path)!
+        let requestHeader = APIHeader.createAuthorizationRequestHeader(userID: userID, authToken: authToken)
+        let requestBody: Request? = requestBody
+        let response: MessageResponse = try await APIClient().request(url: url, httpMethod: .post, requestHeader: requestHeader, requestBody: requestBody)
+    }
+    
+    func deleteAlarm(userID: String, authToken: String, requestBody: AlarmDeleteRequest) async throws {
+        let path = "/alarm/delete"
+        let url = URL(string: environmentVariable.apiEndpoint + path)!
+        let requestHeader = APIHeader.createAuthorizationRequestHeader(userID: userID, authToken: authToken)
+        let requestBody: Request? = requestBody
+        let response: MessageResponse = try await APIClient().request(url: url, httpMethod: .post, requestHeader: requestHeader, requestBody: requestBody)
     }
 }
